+++ oyoyo/client.py
@@ -82,11 +82,34 @@ class IRCClient:
         ...     cli_con.next()
         ...
         """
-        self.context = ssl.create_default_context()
-        self.context.check_hostname = False
-        self.context.verify_mode = ssl.CERT_NONE
-        self.bare_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        self.socket = self.context.wrap_socket(self.bare_socket)
+
+        # This should be moved to profiles
+        import json
+        
+        try:
+            with open("server.json", "r") as server_file:
+                read_file = server_file.read()
+                server_file.close()
+                server_obj = json.loads(read_file)
+            TLS = server_obj['TLS']
+            print("TLS-status is: " + str(TLS))
+            if TLS == False:
+                #print("false")
+                self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+            else:
+                self.context = ssl.create_default_context()
+                self.context.check_hostname = False
+                self.context.verify_mode = ssl.CERT_NONE
+                self.bare_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+                self.socket = self.context.wrap_socket(self.bare_socket)
+        except:
+            print("TLS except.")
+            self.context = ssl.create_default_context()
+            self.context.check_hostname = False
+            self.context.verify_mode = ssl.CERT_NONE
+            self.bare_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+            self.socket = self.context.wrap_socket(self.bare_socket)
+        
         self.nick = None
         self.real_name = None
         self.host = None